Fix Necromancer pets often queueing spells by themselves Fix the casting service sometimes eating up a start cast spell request.

+++ GameServer/ECS-Components/CastingComponent.cs
@@ -60,11 +60,8 @@ namespace DOL.GS
 
             SpellHandler?.Tick(time);
 
-            if (SpellHandler == null && QueuedSpellHandler == null)
-            {
-                _startCastSpellRequests.Clear();
+            if (SpellHandler == null && QueuedSpellHandler == null && _startCastSpellRequests.Count == 0)
                 EntityManagerId = EntityManager.Remove(EntityManager.EntityType.CastingComponent, EntityManagerId);
-            }
         }
 
         public bool RequestStartCastSpell(Spell spell, SpellLine spellLine, ISpellCastingAbilityHandler spellCastingAbilityHandler = null, GameLiving target = null)
